Buttons

+++ Arduino/Tablero/Functions.ino
@@ -1,6 +1,6 @@
-String getValue(String data, char separator, int index)
+String getValue(String data, char separator, byte index)
 {
-    int found = 0;
+    byte found = 0;
     int strIndex[] = { 0, -1 };
     int maxIndex = data.length();
